Moves most of helper functions to separate modules. Groups modules by view
