--- src/fonts/liberation-sans-bold.css
@@ -1,4 +1,0 @@
-@font-face {
-    font-family: 'LiberationSansBold';
-    src: url('liberation-sans-bold.woff') format('woff'), url('liberation-sans-bold.woff2') format('woff2');
-}
