+++ src/utils/Draw.ts
@@ -1,18 +1,4 @@
-import { DependencyNode, Network } from '../components/types';
-import { handleDrag } from './helpers/GraphHelpers';
-import {
-    createDetailsButton,
-    createHighlightBackground,
-    createLabels,
-    createLinkElements,
-    createLinkPath,
-    createMarkers,
-    createSimulation,
-    createSVGContainer,
-    createTextElements,
-    createTooltip,
-    createZoom,
-} from './helpers/DrawHelpers';
+import { Network } from '../components/types';
 import {
     subscribeToChangeHighlightRangeOnArrowKey,
     subscribeToCloseDetails,
@@ -22,57 +8,18 @@ import {
     subscribeToShowTooltipOnNodeHover,
     subscribeToZoomOnArrowKey,
 } from './helpers/UserEventHelpers';
-import { createDetailsViewContainer } from './helpers/DetailsDrawHelpers';
-import { ElementIds } from './AppConsts';
+import { createDetailsViewContainer } from '../details/details-container';
+import { createOverview } from '../overview/overview.helpers';
 
 export const draw = (network: Network, container: HTMLDivElement) => {
-    const { nodes, links, detailsNodes } = network;
-    const width = container.clientWidth;
-    const height = container.clientHeight;
-
-    const simulation = createSimulation(nodes, links, width, height);
-    const svgContainer = createSVGContainer(width, height);
-    const zoomLayer = createZoom(svgContainer, ElementIds.OVERVIEW_ZOOM);
-
-    createMarkers(svgContainer);
-
-    createHighlightBackground(zoomLayer);
-
-    const labelNodesGroup = zoomLayer.append('g').attr('id', ElementIds.LABELS);
-    const linkElements = createLinkElements(zoomLayer, links);
-
-    createTextElements(labelNodesGroup, nodes);
-    createLabels(labelNodesGroup, nodes);
-    createDetailsButton(zoomLayer);
+    createOverview(container, network);
     createDetailsViewContainer(container.clientWidth, container.clientHeight);
-    createTooltip(zoomLayer);
-
-    labelNodesGroup.selectAll<SVGGElement, DependencyNode>('g').call(handleDrag(simulation));
 
     subscribeToHighlight();
     subscribeToResetHighlight();
     subscribeToChangeHighlightRangeOnArrowKey();
     subscribeToZoomOnArrowKey();
-    subscribeToOpenDetails(detailsNodes);
+    subscribeToOpenDetails(network.detailsNodes);
     subscribeToCloseDetails();
     subscribeToShowTooltipOnNodeHover();
-
-    simulation.on('tick', () => {
-        linkElements.each(createLinkPath);
-
-        labelNodesGroup
-            .selectAll('text')
-            .data(nodes)
-            .attr('x', (node: DependencyNode) => (node.x ? node.x : null))
-            .attr('y', (node: DependencyNode) => (node.y ? node.y : null));
-
-        labelNodesGroup
-            .selectAll<Element, DependencyNode>('path')
-            .attr('transform', (node: DependencyNode) =>
-                node.x && node.y ? 'translate(' + (node.x - 30) + ',' + (node.y - 55) + ')' : null
-            )
-            .lower();
-    });
-
-    return svgContainer.node();
 };
